+++ topi/include/topi/reduction.h
@@ -257,7 +257,7 @@ FCommReduce MakeCommReducer(FCombine fcombine,
     auto combiner = tvm::ir::CommReducerNode::make(lhs, rhs, result, id_elem);
     Array<Expr> outputs;
     for (size_t i = 0; i < exprs.size(); ++i) {
-      outputs.push_back(tvm::ir::Reduce::make(combiner, exprs, axis, cond, i));
+      outputs.push_back(tvm::ir::Reduce::make(combiner, exprs, axis, cond, static_cast<int>(i)));
     }
     return outputs;
   };
